% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-in-reference.R
\name{BicsSectorBridge}
\alias{BicsSectorBridge}
\title{Dataset that should probably be removed or outsourced to data store}
\usage{
BicsSectorBridge()
}
\description{
Dataset that should probably be removed or outsourced to data store
}
\examples{
if (dropbox_exists()) {
BicsSectorBridge()
}
}
\seealso{
Other datasets in Reference to remove or outsource to data store: \code{\link{GROUPS.GOVT}},
  \code{\link{RevenueSplit}}, \code{\link{SectorBridge}},
  \code{\link{TYPE.EQUITY}}, \code{\link{TYPE.OTHERS}},
  \code{\link{sector.bridge}}

Other possible_snapshots: \code{\link{ALD.BV}},
  \code{\link{ALD.CB}}, \code{\link{ALD.CC}},
  \code{\link{ALD.EQ}}, \code{\link{ALD.SPV}},
  \code{\link{BALANCE.SHEET.DATA}},
  \code{\link{BENCH.REGIONS}}, \code{\link{CB_OG}},
  \code{\link{DebtMarketClimate}},
  \code{\link{DebtMarket}}, \code{\link{EQMarket.Size}},
  \code{\link{EQ_OG}}, \code{\link{FIN.DATA}},
  \code{\link{Fund.Data}}, \code{\link{FundsTrusts}},
  \code{\link{GROUPS.GOVT}}, \code{\link{INDEX.REGIONS}},
  \code{\link{Indices}}, \code{\link{LoanMarketClimate}},
  \code{\link{LoanMarket}}, \code{\link{PHYSICAL.RISK.CB}},
  \code{\link{PHYSICAL.RISK.EQ}}, \code{\link{Receipts}},
  \code{\link{RevenueSplit}}, \code{\link{SCENLong}},
  \code{\link{SCEN}}, \code{\link{SEC.TYPE.BONDS}},
  \code{\link{SectorBridge}}, \code{\link{TYPE.BONDS}},
  \code{\link{TYPE.EQUITY}}, \code{\link{TYPE.OTHERS}},
  \code{\link{TYPE.RECEIPTS}}, \code{\link{sector.bridge}}
}
\concept{datasets in Reference to remove or outsource to data store}
\concept{possible_snapshots}
