% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_bridge.R
\name{classification_bridge}
\alias{classification_bridge}
\alias{isic_classification}
\alias{nace_classification}
\alias{naics_classification}
\title{Datasets to bridge (translate) common sector-classification codes}
\value{
A \link[dplyr:tibble]{dplyr::tibble}.
}
\description{
These datasets help to translate (bridge) common sector classification codes
from the wild to codes we use in 2dii such as 'power', 'oil and gas', 'coal',
'automotive', 'aviation', 'concrete', 'steel', and 'shipping'.
}
\examples{
isic_classification
nace_classification
naics_classification
}
\seealso{
\code{\link[=data_dictionary]{data_dictionary()}}.
}
\concept{datasets for bridging sector classification codes}
