% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_dictionary.R
\name{create_data_dictionary}
\alias{create_data_dictionary}
\title{Help extend the \code{data_dictionary}}
\usage{
create_data_dictionary(dataset, package = NULL)
}
\arguments{
\item{dataset}{A length-1 character string giving the name of a dataset.}

\item{package}{A length-1 character string giving the name of a package.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
This function takes the exported dataset from a package and creates a
tibble suitable for extending the \code{data_dictionary}.
}
\details{
You may combine \code{create_data_dictionary()} with \code{datapasta::tribble_paste()} to
create a tibble you can conveniently fill. For example,

Code like this:\preformatted{create_data_dictionary("loanbook_demo", package = "r2dii.dataraw") \%>\%
  datapasta::tribble_paste()
}

Outputs something like this:\preformatted{tibble::tribble(
  ~dataset,                                         ~column, ~definition,
  "loanbook_demo",                                "id_loan",          NA,
  "loanbook_demo",                    "id_direct_loantaker",          NA,
...
)
}

You can now replace each \code{NA} with a corresponding \code{definition}.
}
\examples{
create_data_dictionary("loanbook_demo", package = "r2dii.dataraw")
}
\seealso{
\link{data_dictionary}.
}
\keyword{internal}
