% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_description.R
\name{market_description}
\alias{market_description}
\title{Description of a generic equity or debt portfolio dataset}
\description{
A equity or debt market portfolio mapped to all physical assets of all
financial securities issuers within a predefined physical or economic scope
The PACTA methodology is applied to the market data to produce a results file
with the market's relative alignment with different climate scenarios for
technologies in climate-relevant sectors. The market results include the
emissions factors and the absolute and scaled production associated with the
market's exposure to each technology in climate-relevant sectors.
}
\seealso{
Other generic datasets: 
\code{\link{ald_description}},
\code{\link{loanbook_description}}
}
\concept{generic datasets}
