% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_reductions.R
\docType{data}
\name{name_reductions}
\alias{name_reductions}
\title{Name reductions dataset}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 61 rows and 2 columns.}
\usage{
name_reductions
}
\description{
Name reductions dataset
}
\examples{
name_reductions
}
\seealso{
\code{\link[=data_dictionary]{data_dictionary()}}.
}
\keyword{datasets}
