% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_demo.R
\docType{data}
\name{portfolio_demo}
\alias{portfolio_demo}
\title{A market dataset for demonstration}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 9 columns.}
\usage{
portfolio_demo
}
\description{
A market dataset for demonstration
}
\examples{
portfolio_demo
}
\seealso{
\link{data_dictionary} \link{portfolio_description}.

Other demo datasets: 
\code{\link{ald_demo}},
\code{\link{data_dictionary}()},
\code{\link{loanbook_demo}},
\code{\link{market_demo}},
\code{\link{overwrite_demo}}
}
\concept{demo datasets}
\keyword{datasets}
