% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_description.R
\name{portfolio_description}
\alias{portfolio_description}
\alias{equity}
\alias{debt}
\title{Description of a generic equity or debt portfolio result dataset}
\description{
A equity or debt portfolio mapped to the physical assets of financial
securities issuers. The PACTA methodology is applied to the mapped data to
produce a results file with the portfolio's relative alignment with different
climate scenarios for technologies in climate-relevant sectors. The portfolio
results include the emissions factors and the absolute and scaled production
associated with the portfolio's exposure to each technology in
climate-relevant sectors.
}
\seealso{
\link{portfolio_demo}

Other generic datasets: 
\code{\link{ald_description}},
\code{\link{loanbook_description}}
}
\concept{generic datasets}
