% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_classification_df.R
\name{sector_classification_df}
\alias{sector_classification_df}
\title{A view of available sector classification datasets}
\usage{
sector_classification_df()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}. The column \code{code_system} names one of the
classification systems that 2dii uses. All other columns are defined at
\code{\link[=data_dictionary]{data_dictionary()}}.
}
\description{
A view of available sector classification datasets
}
\examples{
sector_classification_df()

data_dictionary() \%>\%
  dplyr::filter(
    grepl("_classification", dataset),
    column \%in\% c("sector", "borderline", "code", "code_system")
 ) \%>\%
 dplyr::arrange(column)
}
\seealso{
\link{nace_classification}, \code{\link[=data_dictionary]{data_dictionary()}}.
}
