% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_snapshot.R
\name{take_snapshot}
\alias{take_snapshot}
\title{Snapshots datasets and the configuration file}
\usage{
take_snapshot(datasets = NULL, destdir = NULL, overwrite = FALSE,
  config = get_config())
}
\arguments{
\item{datasets}{Character vector giving the name of a file (without
extension) in which to write the snapshot. It must exactly match the name
of an exported dataset or of a functions which output is a datasets. \code{NULL}
defaults to take a snapshot of all possible datasets.}

\item{destdir}{Character vector giving the path to the directory to write to.
\code{NULL} defaults to the working directory.}

\item{overwrite}{\code{TRUE} overwrites existing snapshots (data and configuration
file).}

\item{config}{Path to a configuration file.}
}
\value{
Returns \code{NULL}. It's called for its side effect.
}
\description{
Snapshots datasets and the configuration file
}
\examples{
\dontrun{
library(r2dii.utils)

# Use `config` to provide a local, custom configuration file, locally -- to
# affect this specific call to `take_snapshot()`
datasets <- "DebtMarketClimate"
custom_config <- r2dii.utils::example_config("config-toy.yml")
take_snapshot(
  datasets, destdir = tempdir(), overwrite = TRUE, config = custom_config
 )

# Use `options(r2dii_config = <custom_config>)` to provide a custom
# configuration file, globally -- to affect your entire R session
restore <- options(r2dii_config = custom_config)
take_snapshot(
  datasets, destdir = tempdir(), overwrite = TRUE
)

options(restore)
}
}
\seealso{
\link{get_config}.

Other handle snapshots: \code{\link{possible_snapshots}}
}
\concept{handle snapshots}
