% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\alias{create_data_dictionary}
\title{Data dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.}
\usage{
data_dictionary

create_data_dictionary(dataset)
}
\arguments{
\item{dataset}{A dataframe.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
\itemize{
\item \code{data_dictionary} is a dataset that defines the columns of other r2dii
datasets.
\item \code{create_data_dictionary()} is a function that helps extend
\code{data_dictionary}. It is particularly useful when combined with
\code{datapasta::tibble_paste()}. See details.
}
}
\details{
You may combine \code{create_data_dictionary()} with \code{datapasta::tribble_paste()}.
For example, this code:\preformatted{library(r2dii.dataraw)
new_data <- tibble::tibble(x = 1, y = "a")

datapasta::tribble_paste(
  create_data_dictionary(new_data)
)
}

outputs this text so you can conveniently fill the \code{definition} column:\preformatted{tibble::tribble(
  ~dataset,   ~column, ~definition,
  "new_data",     "x",          NA,
  "new_data",     "y",          NA
)
}
}
\examples{
data_dictionary

new_dataset <- tibble::tibble(x = 1, y = "a")
create_data_dictionary(new_dataset)
}
\seealso{
Other demo datasets: \code{\link{loanbook_demo}}
}
\concept{demo datasets}
\keyword{datasets}
