% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.}
\usage{
data_dictionary
}
\description{
This dataset defines the columns of r2dii datasets.
}
\examples{
data_dictionary
}
\seealso{
\link{create_data_dictionary}.

Other generic datasets: \code{\link{loanbook_description}}
}
\concept{generic datasets}
\keyword{datasets}
