% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\title{A dataset that defiles the columns of r2dii datasets.}
\usage{
data_dictionary()
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
A dataset that defiles the columns of r2dii datasets.
}
\examples{
data_dictionary()
}
\seealso{
Other demo datasets: \code{\link{loanbook_demo}}
}
\concept{demo datasets}
